Improved centralization of graphql schema's tests
