+++ src/niweb/apps/noclook/tests/schema/community/__init__.py
@@ -8,59 +8,11 @@ from django.db import connection
 from apps.noclook import helpers
 from apps.noclook.models import NodeHandle, Dropdown, Choice, Role, Group, GroupContextAuthzAction, NodeHandleContext
 from apps.noclook.tests.neo4j_base import NeoTestCase
+from apps.noclook.tests.schema.base import Neo4jGraphQLGenericTest
 
-class TestContext():
-    def __init__(self, user, *ignore):
-        self.user = user
-
-class Neo4jGraphQLTest(NeoTestCase):
+class Neo4jGraphQLTest(Neo4jGraphQLGenericTest):
     def setUp(self):
         super(Neo4jGraphQLTest, self).setUp()
-        self.context = TestContext(self.user)
-
-        # create group for read in community context
-        self.group_read  = Group( name="Group can read the community context" )
-        self.group_read.save()
-
-        # create group for write in community context
-        self.group_write = Group( name="Group can write for the community context" )
-        self.group_write.save()
-
-        # create group for list in community context
-        self.group_list = Group( name="Group can list for the community context" )
-        self.group_list.save()
-
-        # add user to this group
-        self.group_read.user_set.add(self.user)
-        self.group_write.user_set.add(self.user)
-        self.group_list.user_set.add(self.user)
-
-        # get read aa
-        self.get_read_authaction  = sriutils.get_read_authaction()
-        self.get_write_authaction = sriutils.get_write_authaction()
-        self.get_list_authaction  = sriutils.get_list_authaction()
-
-        # get default context
-        self.community_ctxt = sriutils.get_community_context()
-
-        # add contexts and profiles
-        GroupContextAuthzAction(
-            group = self.group_read,
-            authzprofile = self.get_read_authaction,
-            context = self.community_ctxt
-        ).save()
-
-        GroupContextAuthzAction(
-            group = self.group_write,
-            authzprofile = self.get_write_authaction,
-            context = self.community_ctxt
-        ).save()
-
-        GroupContextAuthzAction(
-            group = self.group_list,
-            authzprofile = self.get_list_authaction,
-            context = self.community_ctxt
-        ).save()
 
         # create nodes
         self.organization1 = self.create_node('organization1', 'organization', meta='Logical')
